requirements.txt
Django==2.1.2
Jinja2>=2.9.6
psycopg2
pdf2image
pillow
django-import-export